Refactor the History Table Component.
